pyproject.toml
[project]
name = "ai-customer-support"
version = "0.1.0"
description = "An AI customer support agent using LangChain and LangGraph."
authors = [{ name = "Your Name", email = "your.email@example.com" }]
dependencies = [
    "langchain-core>=0.3.49",           # Updated to match langchain-groq requirement
    "langchain-groq==0.3.2",            # Fixed version
    "langchain-community>=0.2.10",      # Allow newer versions
    "langgraph==0.4.8",
    "python-dotenv==1.0.1",
    "langsmith>=0.1.125,<0.4",
    "tenacity==8.2.0",
    "streamlit==1.32.0"
]
requires-python = ">=3.11"
readme = "README.md"
license = { text = "MIT" }

[tool.uv]
python = ">=3.11"
http-timeout = 180
